add wiki button, move themes to new directory
