+++ frontend/src/components/Navbar.tsx
@@ -5,7 +5,7 @@ import Toolbar from "@mui/material/Toolbar";
 import Button from "@mui/material/Button";
 import LoginWithGithub from "./loginWithGithub";
 import {ThemeProvider} from "@mui/material/styles";
-import {lightTheme} from "../lightTheme";
+import {lightTheme} from "../style/lightTheme";
 
 export default function Navbar() {
 	return (
@@ -16,6 +16,11 @@ export default function Navbar() {
 						<Box sx={{display: "flex"}}>
 							<Button href="/" color="inherit">
 								Home
+							</Button>
+							<Button
+								href="https://github.com/devxp-ca/devxp/wiki"
+								color="inherit">
+								Wiki
 							</Button>
 							<Button href="/wizard" color="inherit">
 								Wizard
